requirements.txt
aiofiles
jinja2
python-multipart
starlette
uvicorn[standard]
yt-dlp
python-slugify
